+++ libraries/ms-common/src/stm32f0xx/spi.c
@@ -26,7 +26,7 @@ StatusCode spi_init(SPIPort spi, const SPISettings *settings) {
   RCC_ClocksTypeDef clocks;
   RCC_GetClocksFreq(&clocks);
 
-  size_t index = __builtin_ffsl(clocks.PCLK_Frequency / settings->baudrate);
+  size_t index = (size_t)__builtin_ffsl((long)(clocks.PCLK_Frequency / settings->baudrate));
   if (index <= 2) {
     return status_msg(STATUS_CODE_INVALID_ARGS, "Invalid baudrate");
   }
@@ -74,7 +74,7 @@ StatusCode spi_exchange(SPIPort spi, uint8_t *tx_data, size_t tx_len, uint8_t *r
                         size_t rx_len) {
   gpio_set_pin_state(&s_port[spi].cs, GPIO_STATE_LOW);
 
-  for (int i = 0; i < tx_len; i++) {
+  for (size_t i = 0; i < tx_len; i++) {
     while (SPI_I2S_GetFlagStatus(s_port[spi].base, SPI_I2S_FLAG_TXE) == RESET) { }
     SPI_SendData8(s_port[spi].base, tx_data[i]);
 
@@ -82,7 +82,7 @@ StatusCode spi_exchange(SPIPort spi, uint8_t *tx_data, size_t tx_len, uint8_t *r
     SPI_ReceiveData8(s_port[spi].base);
   }
 
-  for (int i = 0; i < rx_len; i++) {
+  for (size_t i = 0; i < rx_len; i++) {
     while (SPI_I2S_GetFlagStatus(s_port[spi].base, SPI_I2S_FLAG_TXE) == RESET) { }
     SPI_SendData8(s_port[spi].base, 0x00);
